Aggiunta annotation per verificare la registrazione e test connection su invio mail

+++ src/main/java/com/accenture/interview/controller/base/BaseController.java
@@ -6,14 +6,14 @@ package com.accenture.interview.controller.base;
 public abstract class BaseController {
 
 	/** The interview id. */
-	protected static Integer interviewId;
+	protected static Long interviewId;
 
 	/**
 	 * Update interview id.
 	 *
 	 * @param value the value
 	 */
-	protected static void updateInterviewId(Integer value) {
+	protected static void updateInterviewId(Long value) {
 		interviewId = value;
 	}
